Effects.FileSystem.MonadPathReader

Description

Provides the MonadPathReader effect.

Since: 0.1

Class

class Monad m => MonadPathReader m where

Represents file-system reader effects.

Since: 0.1

Methods

listDirectory :: HasCallStack => Path -> m [Path]

Since: 0.1

getDirectoryContents :: HasCallStack => Path -> m [Path]

Since: 0.1

getCurrentDirectory :: HasCallStack => m Path

Since: 0.1

getHomeDirectory :: HasCallStack => m Path

Since: 0.1

getXdgDirectory :: HasCallStack => XdgDirectory -> Path -> m Path

Since: 0.1

getXdgDirectoryList :: HasCallStack => XdgDirectoryList -> m [Path]

Since: 0.1

getAppUserDataDirectory :: HasCallStack => Path -> m Path

Since: 0.1

getUserDocumentsDirectory :: HasCallStack => m Path

Since: 0.1

getTemporaryDirectory :: HasCallStack => m Path

Since: 0.1

getFileSize :: HasCallStack => Path -> m Integer

Since: 0.1

canonicalizePath :: HasCallStack => Path -> m Path

Since: 0.1

makeAbsolute :: HasCallStack => Path -> m Path

Since: 0.1

makeRelativeToCurrentDirectory :: HasCallStack => Path -> m Path

Since: 0.1

doesPathExist :: HasCallStack => Path -> m Bool

Since: 0.1

doesFileExist :: HasCallStack => Path -> m Bool

Since: 0.1

doesDirectoryExist :: HasCallStack => Path -> m Bool

Since: 0.1

findExecutable :: HasCallStack => String -> m (Maybe Path)

Since: 0.1

findExecutables :: HasCallStack => String -> m [Path]

Since: 0.1

findExecutablesInDirectories :: HasCallStack => [Path] -> String -> m [Path]

Since: 0.1

findFile :: HasCallStack => [Path] -> String -> m (Maybe Path)

Since: 0.1

findFiles :: HasCallStack => [Path] -> String -> m [Path]

Since: 0.1

findFileWith :: HasCallStack => (Path -> m Bool) -> [Path] -> String -> m (Maybe Path)

Since: 0.1

findFilesWith :: HasCallStack => (Path -> m Bool) -> [Path] -> String -> m [Path]

Since: 0.1

pathIsSymbolicLink :: HasCallStack => Path -> m Bool

Since: 0.1

getSymbolicLinkTarget :: HasCallStack => Path -> m Path

Since: 0.1

getPermissions :: HasCallStack => Path -> m Permissions

Since: 0.1

getAccessTime :: HasCallStack => Path -> m UTCTime

Since: 0.1

getModificationTime :: HasCallStack => Path -> m UTCTime

Since: 0.1

type Path = FilePath

For filepath >= 1.4.100 and directory >= 1.3.8, Path = System.FilePath.OsPath. Below that it is a FilePath.

Since: 0.1

Xdg Utils

getXdgConfig :: (HasCallStack, MonadPathReader m) => Path -> m Path

Retrieves the Xdg Config directory.

Since: 0.1

Reexports

data XdgDirectory

Special directories for storing user-specific application data, configuration, and cache files, as specified by the XDG Base Directory Specification.

Note: On Windows, XdgData and XdgConfig usually map to the same directory.

Since: directory-1.2.3.0

Constructors

XdgData
For data files (e.g. images). It uses the XDG_DATA_HOME environment variable. On non-Windows systems, the default is ~/.local/share. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /usr/share.

XdgConfig
For configuration files. It uses the XDG_CONFIG_HOME environment variable. On non-Windows systems, the default is ~/.config. On Windows, the default is %APPDATA% (e.g. C:/Users/<user>/AppData/Roaming). Can be considered as the user-specific equivalent of /etc.

XdgCache
For non-essential files (e.g. cache). It uses the XDG_CACHE_HOME environment variable. On non-Windows systems, the default is ~/.cache. On Windows, the default is %LOCALAPPDATA% (e.g. C:/Users/<user>/AppData/Local). Can be considered as the user-specific equivalent of /var/cache.

data XdgDirectoryList

Search paths for various application data, as specified by the XDG Base Directory Specification.

The list of paths is split using searchPathSeparator, which on Windows is a semicolon.

Note: On Windows, XdgDataDirs and XdgConfigDirs usually yield the same result.

Since: directory-1.3.2.0

Constructors

XdgDataDirs
For data files (e.g. images). It uses the XDG_DATA_DIRS environment variable. On non-Windows systems, the default is /usr/local/share/ and /usr/share/. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

XdgConfigDirs
For configuration files. It uses the XDG_CONFIG_DIRS environment variable. On non-Windows systems, the default is /etc/xdg. On Windows, the default is %PROGRAMDATA% or %ALLUSERSPROFILE% (e.g. C:/ProgramData).

data Permissions

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)