Effects.FileSystem.MonadPathWriter

Description

Provides the MonadPathWriter effect.

Since: 0.1

Class

class Monad m => MonadPathWriter m where

Represents file-system writer effects.

Since: 0.1

Methods

createDirectory :: HasCallStack => Path -> m ()

Since: 0.1

createDirectoryIfMissing :: HasCallStack => Bool -> Path -> m ()

Since: 0.1

removeDirectory :: HasCallStack => Path -> m ()

Since: 0.1

removeDirectoryRecursive :: HasCallStack => Path -> m ()

Since: 0.1

removePathForcibly :: HasCallStack => Path -> m ()

Since: 0.1

renameDirectory :: HasCallStack => Path -> Path -> m ()

Since: 0.1

setCurrentDirectory :: HasCallStack => Path -> m ()

Since: 0.1

withCurrentDirectory :: HasCallStack => Path -> m a -> m a

Since: 0.1

removeFile :: HasCallStack => Path -> m ()

Since: 0.1

renameFile :: HasCallStack => Path -> Path -> m ()

Since: 0.1

renamePath :: HasCallStack => Path -> Path -> m ()

Since: 0.1

copyFile :: HasCallStack => Path -> Path -> m ()

Since: 0.1

copyFileWithMetadata :: HasCallStack => Path -> Path -> m ()

Since: 0.1

createFileLink :: HasCallStack => Path -> Path -> m ()

Since: 0.1

createDirectoryLink :: HasCallStack => Path -> Path -> m ()

Since: 0.1

removeDirectoryLink :: HasCallStack => Path -> m ()

Since: 0.1

setPermissions :: HasCallStack => Path -> Permissions -> m ()

Since: 0.1

copyPermissions :: HasCallStack => Path -> Path -> m ()

Since: 0.1

setAccessTime :: HasCallStack => Path -> UTCTime -> m ()

Since: 0.1

setModificationTime :: HasCallStack => Path -> UTCTime -> m ()

Since: 0.1

type Path = FilePath

For filepath >= 1.4.100 and directory >= 1.3.8, Path = System.FilePath.OsPath. Below that it is a FilePath.

Since: 0.1

Misc

removeFileIfExists :: (HasCallStack, MonadPathReader m, MonadPathWriter m) => Path -> m ()

Calls removeFile if doesFileExist is True.

Since: 0.1

removeDirectoryIfExists :: (HasCallStack, MonadPathReader m, MonadPathWriter m) => Path -> m ()

Calls removeDirectory if doesDirectoryExist is True.

Since: 0.1

removeDirectoryRecursiveIfExists :: (HasCallStack, MonadPathReader m, MonadPathWriter m) => Path -> m ()

Calls removeDirectoryRecursive if doesDirectoryExist is True.

Since: 0.1

removePathForciblyIfExists :: (HasCallStack, MonadPathReader m, MonadPathWriter m) => Path -> m ()

Calls removePathForcibly if doesPathExist is True.

Since: 0.1

Reexports

data Permissions

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)